refactor: HttpRequestUrl에서 HttpRequestUrlParameters 분리
